Modifiy session type to Aqua; trim plist
I was having trouble with the agent staying loaded after a reboot
or logout/login. I still do not understand why it was happening.
But I changed the session type to Aqua, which means (I think) that
the agent gets loaded when the user logs in interactively. I
don't know when Background gets loaded or when System session
types get loaded.

I also trimmed unnecessary keywords from the plist. There may
be more to trim.