Use gnetcat to detect unresponsive hosts
The script was hanging if a host was unreachable or unresponsive.
I added a check using gnetcat that will timeout after 2 seconds if
host doesn't respond ... validation for unresponive hosts is skipped
(obviously).

This only works with gnetcat. The version of netcat that is installed
on macOS (10.12.6 and earlier; and maybe later) ignores the '-w'
argument, so will instead hang for a couple minutes waiting for the
host to respond. So I test for gnetcat and use it; if not avialable
then we skip host validation and go right to the openssl test, which
may hang on an unresponsive host.